refactor(1013): remove DisplayUtils facade (SC-001, position 11)
Extract DisplayUtils (72 LOC, table + progress-bar helpers) from MistHelper.py
to src/ui/display_utils.py per initiative 1013 Cat B position 11.

- Add lazy MistHelper import for DataProcessingUtils.get_unique_keys
- Add re-export in MistHelper.py for existing lazy _MH.DisplayUtils callers
  (src/firmware/firmware_manager.py x4, src/refactors/device_data_fetcher.py x1)
- Delete class body, leave breadcrumb pointer

Gates: black, ruff, py_compile, vulture, --test 66/0/131 (0 failed)
